LPS-96399 - Move testSearchIncludeAttachment to WikiAttachmentsTest
