+++ modules/apps/document-library/document-library-test/build.gradle
@@ -28,7 +28,6 @@ dependencies {
 	testIntegrationCompile project(":apps:trash:trash-api")
 	testIntegrationCompile project(":apps:trash:trash-test-util")
 	testIntegrationCompile project(":apps:users-admin:users-admin-test-util")
-	testIntegrationCompile project(":apps:wiki:wiki-api")
 	testIntegrationCompile project(":core:petra:petra-function")
 	testIntegrationCompile project(":core:petra:petra-reflect")
 	testIntegrationCompile project(":core:petra:petra-string")
